+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - TBD
+### Changed
+- Removed `rootFileName` from calls to `USE_F`, `USE_F90`, `USE_C`, `USE_MATLAB` to match the function prototypes in `code.h`
+
 ## [3.1.1] - 2024-04-30
 ### Changed
 - Updated Python package versions for ReadTheDocs in `docs/requirements.txt`